+++ R/auk-rollup.R
@@ -77,7 +77,7 @@
 #' 
 #' # yellow-rump warbler subspecies rollup
 #' library(dplyr)
-#' # without rollup, there are three observations
+#' # without rollup, there are multiple observations per checklists
 #' ebd |>
 #'   filter(common_name == "Yellow-rumped Warbler") |>
 #'   select(checklist_id, category, common_name, subspecies_common_name,