+++ NEWS.md
@@ -1,6 +1,7 @@
 # auk 0.9.1
 
 - ensure taxon_concept_id behaves correctly in auk_rollup() (issue #94)
+- update EBD example files to get latest format (e.g. add taxon_concept_id)
 
 # auk 0.9.0